split conflict detector in parts, indexer with cache
